final entry setup with modular graphx and union find

+++ src/main/java/spark/PartitionConfiguration.java
@@ -4,12 +4,12 @@ import java.io.Serializable;
 
 public class PartitionConfiguration implements Serializable {
 
-        final double minLatitude, maxLatitude, minLongitude, maxLongitude;
-        final double cellSize;
-        final double buffer;
+        final float minLatitude, maxLatitude, minLongitude, maxLongitude;
+        final float cellSize;
+        final float buffer;
         final int numCellsX, numCellsY;
 
-        PartitionConfiguration(double minX, double maxX, double minY, double maxY, double eps) {
+        PartitionConfiguration(float minX, float maxX, float minY, float maxY, float eps) {
             this.minLatitude = minX;
             this.maxLatitude = maxX;
             this.minLongitude = minY;
